Reorganise components
This puts "page" components into `pages`, and everything else into
`components`.
