+++ src/components/ReagentCheckbox.jsx
